Add files via upload
The code of these files are not intended to be neither efficient nor optimum. Also the results shown by the algorithms are not accurate at all. The code are only for academic purposes and, at any time, it's been trying to get good results or performances. It can be seen that in, for instance, at any time any feature engineering was made, even the most common, as well any effort was made in tunning the models.